requirements.txt
veracode-api-py>=0.9.62
logging-formatter-anticrlf>=1.2